# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MCP Forge API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/projects/{project_id}/builds": {
      "post": {
        "tags": [
          "Builds"
        ],
        "summary": "Trigger Build",
        "description": "Trigger a new build for the project.",
        "operationId": "trigger_build_api_projects__project_id__builds_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuildSummaryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Builds"
        ],
        "summary": "List Project Builds",
        "description": "List all builds for a project.",
        "operationId": "list_project_builds_api_projects__project_id__builds_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BuildSummaryResponse"
                  },
                  "title": "Response List Project Builds Api Projects  Project Id  Builds Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/builds/{build_id}": {
      "get": {
        "tags": [
          "Builds"
        ],
        "summary": "Get Build Summary",
        "description": "Get single build metadata.",
        "operationId": "get_build_summary_api_builds__build_id__get",
        "parameters": [
          {
            "name": "build_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Build Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuildSummaryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/builds/{build_id}/files": {
      "get": {
        "tags": [
          "Builds"
        ],
        "summary": "Get Files Tree",
        "description": "Get hierarchical file tree of generated project.",
        "operationId": "get_files_tree_api_builds__build_id__files_get",
        "parameters": [
          {
            "name": "build_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Build Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Get Files Tree Api Builds  Build Id  Files Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/builds/{build_id}/files/content": {
      "get": {
        "tags": [
          "Builds"
        ],
        "summary": "Get File Content",
        "description": "Retrieve text content of a generated project file safely.",
        "operationId": "get_file_content_api_builds__build_id__files_content_get",
        "parameters": [
          {
            "name": "build_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Build Id"
            }
          },
          {
            "name": "path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Relative file path within generated project",
              "title": "Path"
            },
            "description": "Relative file path within generated project"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileContentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/builds/{build_id}/download": {
      "get": {
        "tags": [
          "Builds"
        ],
        "summary": "Download Build Archive",
        "description": "Download the deterministic ZIP artifact for a build.",
        "operationId": "download_build_archive_api_builds__build_id__download_get",
        "parameters": [
          {
            "name": "build_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Build Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/builds/{build_id}/connect": {
      "get": {
        "tags": [
          "Builds"
        ],
        "summary": "Get Connect Snippets",
        "description": "Get ready-to-copy client configuration snippets (Claude Desktop, Cursor, stdio, HTTP).",
        "operationId": "get_connect_snippets_api_builds__build_id__connect_get",
        "parameters": [
          {
            "name": "build_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Build Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Connect Snippets Api Builds  Build Id  Connect Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "BuildSummaryResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "spec_version_id": {
            "type": "string",
            "title": "Spec Version Id"
          },
          "build_no": {
            "type": "integer",
            "title": "Build No"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "tool_count": {
            "type": "integer",
            "title": "Tool Count"
          },
          "warning_count": {
            "type": "integer",
            "title": "Warning Count"
          },
          "artifact_sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artifact Sha256"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "spec_version_id",
          "build_no",
          "status",
          "tool_count",
          "warning_count",
          "created_at"
        ],
        "title": "BuildSummaryResponse"
      },
...
      "FileContentResponse": {
        "properties": {
          "path": {
            "type": "string",
            "title": "Path"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          }
        },
        "type": "object",
        "required": [
          "path",
          "content",
          "size"
        ],
        "title": "FileContentResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}